Logic for province dropdown (6.3.5)
** public/js/form-submission-helpers.js **
- This was formerly named 'yikes-update-existing-subscriber.js'. I added
a function to it that's not related to updating an existing subscriber
so I renamed it for clarity. This script is enqueued on every page our
shortcode is used on, and can serve as a central place for form-related
JavaScript code in the future.
- Added the onchange country code function that was formerly in
process_form_shortcode

** public/partials/shortcodes/templates/province-dropdown.php **
- HTML for a provinces dropdown in the same manner as the current states
dropdown

** public/partials/process_form_shortcode.php **
- Changed the enqueue'd JavaScript file
- Added filter/logic to control showing the state dropdown
(yikes-mailchimp-display-state-field)
- Added filter/logic to control what we show in the state dropdown field
(yikes-mailchimp-state-field-dropdown and
yikes-mailchimp-custom-state-field-html)
- Added a filter to the default zip code value
- Removed the checkCountry() JS function from the PHP file and placed it
into the new JS file